add index tests
